+++ src/main/java/com/github/ilim/backend/dto/QuizDto.java
@@ -7,6 +7,7 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
@@ -22,7 +23,7 @@ public class QuizDto {
     private String description;
 
     @NotNull
-    private Integer passingScore;
+    private BigDecimal passingScore;
 
     @NotNull
     private List<QuestionDto> questions;
